requirements.txt
python-dotenv
SQLAlchemy
psycopg2-binary
numpy==1.26.4
pandas
tqdm
tf-keras
sentence-transformers
pgvector
asyncio
asyncpg